pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lumina"
version = "0.1.0"
description = "A high-performance custom image manipulation engine."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21.0",
    "Pillow>=9.0.0",
]

[project.scripts]
lumina-cli = "lumina.cli.main:entry_point"